work-queue: count work queue size correctly
works_count is used by the rest of the code as indicator for
number of tasks in the work queue (for example see
Fp-utils.c::print_work_count()), but this variable does not
reflect this value accurately because it is not decremented
immediately on task removal.

Move it to the task removal section so that work queue size is
properly maintained.

Signed-off-by: Nikita Kiryanov <nikita@compulab.co.il>