screen_saver: fix incorrect timeout bug
It is currently impossible to reliably set the screen saver to
the exact timeout requested because the rtc_scheduler works at
1 sec resolution, and so even if a task timeout is reset, the
length of the first "second" is very likely to be less than a
second, unless the timeout reset happened exactly at the point
of RTC interrupt.

Fix this by moving the screen saver task to tc_scheduler, which
is not sensitive to such phenomena.

Signed-off-by: Nikita Kiryanov <nikita@compulab.co.il>