layout: provide non-misleading default defines
The defaults for VERSION, BUILD_DATE, and BUILD_TIME, lie about their
own values. The defaults
should tell the user that this binary was compiled without proper
version/date/time values, but you
wouldn't be able to tell this with the defaults that are currently in
use.

Provide more telling defaults.