clock: fix blinking seconds garbage strings
The clock screen saver is supposed to have a blinking colon separating
hours from the minutes. Quite often, instead of the colon blinking on
and off, it blinks between a colon and some garbage value. This is
caused by an incorrect assignment of the empty string in the rtc seconds
handler function.

Fix it by using sprintf.