Fp-utils: get rid of check_update_x_req() functions
We have a function called update_requests(). One would expect it to
update requests, and it does. But it does so using multiple functions
that are called check_update_x_req(), which are named as if they are
only checking for updates, while in practice they also set the values to
"updated".

This whole thing is confusing and unintuitive. Get rid of the
check_update_..() funcitons, which are the main cause for confusion.