get rid of buggy set_curr_str()
set_curr_str is completely broken. It's supposed to be used to obtain
the string representation of a current value, but it fails to do that by
overwriting the destination pointer with a pointer to an arbitrary empty
string. The result is that not only does the intended destination not
get written, what does gets written is some random memory location,
which can easily cause fatal state corruption.

This can be easily fixed. However, when we look at how the function is
used, we see that it's only used in functions that attempt to determine
if a value needs to be updated. The bug in set_curr_str() causes that
part of the check to almost always come up true, and the front panel app
works perfectly well with that MO, so this is an opportunity to simplify
the decision logic by deleting all the set_curr_str() dependant segments
and just use true instead.