gfx: utils: refactor control sign drawing to remove is_numric
draw_graphic_signs accepts a is_numeric parameter, which essentially
causes it to handle two cases instead of focusing on one. This is not
the proper way to go. Instead of hardcoding special cases, rewrite it
such that it accepts the data it's supposed to print, and the selection
of data is done by the affected frame data. This makes the gfx code more
generic and encapsulates the non-generic decision in the module that
actually deals with the specifics of the display contents (screen code).