gfx_information: introduce draw_controls
Introduce a function to draw symbols to hint at the meaning of the right
and left buttons.
This is an gfx_information function because the symbols depend on the
information value and internal implementation.

This eliminates the need for insert_graphic_signs and for exporting the
set_x_draw_graphic_signs() functions.