gfx: gfx_components: gfx_information: add handle_buttons function
Add handle_buttons() function to the gfx_information object.
case specific handle_button funcitons are dependant on the information
type, and are already implemented in information code. The next logical
step is to standardize that as part of the gfx_information API.

This commit rewrites part of the button handling algorithm to use the
new API function. This rewrite changes the check that is used to invoke
the handle_buttons() function. The change is based on the following
facts:

frame_present is not NULL if we are viewing a frame, and NULL otherwise.
is handle_buttons() is to be invoked, we use the information object at
the list head of the currently active frame (frame_present). This is
because despite what the definition of struct gfx_frame appears to
suggest, every frame has exactly one gfx_information element in
practice, not a list of multiple information objects. The current
algorithm design actually assumes this, so this patch takes advantage of
this. This is a design flaw though, and will be fixed in the future.