gfx: eliminate a check for is_active_frame
The use of is_active_frame is an aberration. Instead of letting every
information object decide if it needs a full redraw or not, the gfx
subsystem tests a flag, which all screens are supposed to set when
displayed, and makes a decision based on that. This is not only cryptic,
but is also hard to maintain.

We are still not ready to fully get rid of it, but here we take one step
closer by using the redraw parameter from information objects that
represent action frames.