gfx: action_menu: get rid of is_active_frame
Currently the only role is_active_frame plays is when an "active frame"
(a frame that performs an action rather than simply displaying
information) is drawn. It is used to set display_state to
DISPLAY_ACTION_FRAME to distinguish it from a regular DISPLAY_FRAME.
Luckily, this state setting is literally the only place where this
distinction is made. Everywhere else in the code DISPLAY_ACTION_FRAMEs
and DISPLAY_FRAMEs are handled the same way, so we can stop tracking the
difference, considering that a DISPLAY_ACTION_FRAME is-a DISPLAY_FRAME
anyway.