gfx: action_menu: get rid if is_graphic_view
The method of keeping track of flags and then give items with special
flags special attention is hard to maintain and complicates the code
unnecessary. For this reason, it's better to use the method in which
special behavior is implemented by the behvaior subject, rather than by
the main algorithm. Therefore, it is our goal to get rid of these flags.

One such flag is is_graphic_view, which differentiates the I3M standard
menu (with icons) from the ASF standard menu (arrow and string list). We
do not currently use ASF style menus, so currently there's no need for
this flag. Once we do support ASF style menus, this support will be
implemented OOP style.

Remove is_graphic_view.