gfx: action_menu: get rid of visible field
The visible field is yet another one of those flags that the algorithm
has to keep track of instead of using a more encapsulated approach to
dispatching behavioral differences. As it happens, we don't actually
need this field right now. frame_present can fill in for it with no
issues.

Get rid of it to simplify the implementation.