gfx: frame: introduce frame centric controls
Until now the responsibility of responding to buttons, as well as
drawing appropriate control symbols, was solely of gfx_information.
However, this method of operation does not work if we need to have a
frame with multiple gfx_information objects, such as in the case of
dashboard screens like clock and dashboard.

The new method of operation is now as follows: the move-to-next-frame
functionality is implemented by default on the frame object. To match
the action to the on-screen control symbols, frames now also have a
draw_controls() function. These sets of functions are used unless
there's a gfx_information specific implementaiton, in which case the
information functions are used.

This allows us to provide default actions that makes sense for all
frames, while still allowing information screens to take control when
the screen is really doing something active.