gfx: action_menu: graphic_menu: stop using is_back_item
The functions that handle cursor movement and icon position are giving
special treatment to the back icon. This is wrong. The functions should
be generic, and the special treatment should be implemented on a higher
level (the one who calls these functions should decide where to place
icons and move cursor to).

Rewrite the functions to not use this parameter anymore.