gfx: action_menu: graphic_menu: introduce new select/deselect functions
Currently the function that handles cursor select/deselect is called
invert_item(), whose correctness of operation depends on past
invocations being properly in sync with each other. This is problematic,
because the correctness of a funciton should depend only on its code,
not on the context of invocation. Introduce new select/deselect
funcitons that only handle either a selection or a deselection, in a
preparation to stop using invert_item.