gfx: action_menu: graphic_menu: redesign graphic_menu_init
Currently graphic_menu_init is being used for drawing the menu, and
moving the cursor.
Ironically, initializing the menu is not something it actually does. So,
rename it to graphic_menu_draw() and remove the handling of the cursor.
Instead, use the graphic menu select/deselect functions.

This causes a cascade effect on the call stack that renders some helper
functions like show_menu and show_current_menu unnecessary anymore.