scheduler: update_screen_task: don't check if frame need update
The is_frame_need_update is a patch function for the slow refresh
problem of some display redrawing situations. However, in this case we
don't redraw the frames, so the check is unnecessary. Remove it so that
it'll be easier to get rid of is_information_need_to_change() someday in
the future.