display: fix blinking menu screen bug
When we have the situation of a refreshing menu screen, the refresh is
very noticeable and jarring. This is because the screen drawing
algorithm for menus uses gfx_mono_generic_put_bitmap() to draw icons in
the framebuffer, but this function also flashes the framebuffer to the
screen. The result is that screen drawing involves multiple screen
refreshes, as opposed to just one.

Deciding when we are ready to flush the framebuffer to the screen should
be our prerogative, not that of gfx_mono_generic_put_bitmap(). Remove
the call to gfx_mono_put_framebuffer().