gfx: frame: get rid of the redraw parameter
The need for a redraw parameter is questionable. The need to avoid a
full redraw can be due to performance constraints, but we don't have
such issues, nor do we have a need to cleanup the screen of no longer
needed artefacts. So, get rid of this parameter and always do a full
redraw. To avoid the blinking screen problem that we used to have in
menus, make prevent glcd_write from refreshing the screen. It should be
our decision when to do a screen refrsh, not that of the display
library.