Introduce a new convention for ordering includes
Over time includes pick up inter-dependencies which are hard to track
down. The new method for ordering includes attempts to minimize this by
going from local to global. i.e.:

1) h file corresponding to this c file
2) headers from the same component
3) headers from other components
4) system headers

Following this order, we prove that each header, for which there's a c
file, can be included without prerequisites, and the rest follows
logically from there.

Minor adjustments to include file contents were made so that Fp-utils.h
could finally be eliminated as part of the reordering.