Get rid of debug.h
The debug.h file contains a set of direct-to-screen printing functions.
As such it is not really a debug header, but a printing function
library, and it is one that isn't being used. Though the functions can
be useful, we don't need them right now and can reintroduce them later
if the need arises. For now, get rid of this header.