gfx: get rid of DMI stuff
The DMI implementation is currently not used in production, and the implementation
itself is really questionable because it replicates standard gfx functionality just for the
sake of content specific modifications, which should be taken care of elsewhere, not in
the gfx library.

For now just get rid of DMI code altogether. In the future we can reimplement it
correctly if the need arise.