gfx: implement distinct types of frames
Up until now a frame was a frame was a frame. This makes defining a usable API
difficult, and also forces gfx_frame_init() to take on multiple responsibilities to handle
all the different types of frames.

This commit creates a hierarchy of frame types. There's the standard gfx_frame_init()
that creates a simple frame that does nothing other than display its contents.
This is the most context versatile frame, because it has no restrictions on when it can
be used. It can be a part of a menu, or it can be spontaneously invoked by the app in
response to various events.

We extend this functional template with gfx_context_frame_init(), which is a
functional frame whose functionality is that it exists in the context of a menu, and so
the hardware controls can be used to move between frames in the same context.
It does not make sense to invoke this frame outside of a menu context.

Another variation on the vanilla gfx_frame is the gfx_action_frame_init(), which is a
type of frame that performs an action. Examples include setting the brightness or
setting screen saver timeout. Like the vanilla frame, it can be invoked out of menu
context.

The result of this refactor, aside from a more semantic API, is that frame initialization
no longer need to accept a dashboard parameter, which was an ugly hack to tack on
a functionality the API wasn't ready for.