gfx_frame: move config block loading to menu-handler
gfx code should not concern itself with the specifics of loading stuff from the config
section. Refactor the code such that all of those defails are taken care of in the file
that handles the vast majority of config section loading: menu-handler.c. In the future,
menu-handler.c should be moved out of gfx folder and into the config section.