gfx: action_menu: move special frame pointers to menu-handler
menu-handler is the one that handles loading config data and using it to initialize
frames, so it is the file that should have the pointers for storing the results.

Move dashbaord, clock, and splash pointers to menu-handler, and make them static
because they are really internal menu-handler data. The code isn't quite ready yet for
this change, so comment out the call to show_logo() until it is refactored away.