config_block: initialize splash using gfx_frame_init()
Splash is currently using a custom draw() function, even though the normal frame
draw function should be able to serve it perfectly. Rewrite the splash init code so that
splash no longer needs a special draw() function. Unfortunately now it needs a special
init() function, but that will be taken care of later.