gfx: move most of action_menu to gfx_gui_control
A lot of the code in gfx_action_menu actually implements the gui state machine and
the funcitons that move between the states.

Move this code to gfx_gui_control, so that gfx_aciton_menu could be all about the
menu.

The function displaying action_menu is slightly modified for the move.