screens: get rid of is_valid implementations
is_valid() is used by the menu display code to mark icons in a special way to designate
features that are not active, or data points that don't have valid data yet.

Make it all about inactive features. Data may not be valid for a variety of reasons, and
there's no utility in communicating to the user that the associated feature is not ready.