screens: implement custom messages as standard frames
Currently, printing a custom message means acting outside of the structure of
the gui framework. Such a practice requires maintaining a bunch of ad-hoc code to
cover for our violation of the navigation model (any change in view should be either
a standard switch to a frame or a standard switch to a menu, but the custom message
is neither).

Implement the custom message using a dedicated frame provided by the
gui_control, and display it using standard switch to frame. This allows us to finally get
rid of the display_state enum. In the process of this reimplementation, information
init has been expanded to allow initializing information with arbitrary data (not from
config block).