config_block: get rid of set_menu_by_id()
set_menu_by_id() is a function that couples config_block loading code with gui state
code, which really shouldn't be necessary.

Eliminate this function by redefining gfx_gui_init() to be just about initialization,
without attempting to display the menu (the rest of the gui subsystem will take care
of that naturally). As part of this redefinition, it will accept the menu definitions,
which is a reasonable parameter, and so there will no longer be a need to callback to
config_block code.