Integrate fp-config project to fp-application
We wish to be able to define a special config section that is separate from the
application itself, so that custom config sections could be installed without changing
the application itself. Because of this requierment, config section was initially defined
in its own project. This created some issues, in particular the need to synchronize a
bunch of enums and structs between the projects, because a copy of these data
types had to be defined in both projects and had to be compatible.

Import the fp-config into fp-application, merging the common data type definitions.
The project setup is updated such that separate compilation of the config section
could be achieved using a special build target. For convenience, fp-application
itself is compiled with the config section already included, and placed in the
appropriate offset as if it was flashed there using the old method of a separate
config hex file.

With this commit, fp-application.cproj is now included in the source code, since it
contains important definitions for the build targets and their setup.