Debug: make compiler flags similar to Release setup
This project had a repeating cycle:
Developemnt is done on Debug build. Project is switched to Release build. Suddenlt, a
bunch of warnings pop up, because Release build is setup to be more pedantic.

Make compiler flags the same in both configuration, so that we would have less of
those "squash warnings" commits.