twi: get rid of twi_enable/disable functions
twi_enable(0 and twi_disable() are not used externally. Make the use of the flag
twi_enable internal, and get rid of the functions since they are not really necessary
(we never disable twi once enabled, and twi_enable is trivial, so there's no added
benefit in the function abstraction.