scheduler: generalize scheduler with regards to TC
Move TC initialization outside of scheduler, and make scheduler accept as parameter
the pointer to the TC that will clock its operation. This opens the way for future
extensions of the scheduler, and also moves the tc init code to the same place where
all the hardware initialization is happening.