ASF: update includes such that files will be compatible with latest ASF
The original version of this app diverged from ASF mode of operation by not using the
src folder ASF is designed to work with. Further diversion was added when the include
paths were changed in order to avoid having to track include paths in the project
configuration. Both were the wrong thing to do, since that makes updating ASF code
more difficult.

Take a step to rectify this by fixing the includes. We're not going to bring back the
src folder just yet, because that will mess up the entire codebase history log.