ASF: gfx_mono_menu: return code to unmodified state
We prefer to not modify ASF libraries so that upgrading them would be easy, and so
that code could be easily moved around. Most of the functions in gfx_mono_menu()
aren't being used. Copy the only one that is in use, with all the modifications, and thus
bring the gfx_mono_menu files to their original state.