sram: do not clear requests
The method of clearing requests after a piece of data has been received assumes one
piece of data per screen. This assumption is no longer true, and since we can only
post requests for entire groups of data (such as, "give me all cpu temps"), we can't
just clear this after a piece of information for one member of the group has been
received. This will likely starve the other members of the group, because as soon as
the first one gets its data, the request is cleared.

Implement a different mechanism: instead of clearing requests on receiving,
information. let the request posting task handle the clearing. This is essentially what
this task already does. If a type of information is not needed, the request task shall
mark it as unnecssary.