Get rid of final remnants of Linux build files
Currently we only support building on Windows. Previously there was
support for Linux builds, but once development moved on to a Windows
environment Linux build support was neglected, and the build files
were not kept up to date. Most of these files have since been removed.

This commit removes the last remnants of the Linux build metadata.