power: update power calculation for Airtop 3
The Airtop3 will support higher power components then previous Airtop models.
In order to maintain components reliability and to allow the power measurement of
the I3M to function properly, the current sense resistors in the Airtop3 mother board
will be changed.
The resistance of the resistors will be lowered in a factor of 1.5, so in order to
compensate it, multiply the power shown in the power measurement screen by 1.5.
Now the I3M power screen will show any value between 6W and to 360W.